pyproject.toml
[project]
name = "felix-case-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.20.0",
    "pyyaml>=6.0.3",
]
